Unify corpus survey logic under analysis.corpus package
