+++ lcats/lcats/analysis/corpus/detectors/__init__.py
@@ -1,0 +1,7 @@
+"""Corpus detector modules."""
+
+from lcats.analysis.corpus.detectors import boundary
+from lcats.analysis.corpus.detectors import structural
+from lcats.analysis.corpus.detectors import unicode
+
+__all__ = ["boundary", "structural", "unicode"]
