+++ lcats/lcats/analysis/corpus_survey.py
@@ -1,1031 +1,69 @@
-"""Corpus survey architecture primitives and detector orchestration."""
-
-import argparse
-import csv
-from dataclasses import dataclass
-import json
-import pathlib
-import re
-import subprocess
-import sys
-from typing import Any, Mapping, Optional, Protocol, Sequence
-import unicodedata
-
-import lcats.inspect
-import tqdm
-
-
-SMART_ALLOWED = {"–", "—", "‘", "’", "“", "”", "…"}
-ASCII_PUNCT = {chr(i) for i in range(32, 127)} - set(
-    "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789"
-)
-DEFAULT_EXCLUDED_CODEPOINTS = [
-    "1F4D6",  # Open Book
-    "1F4DC",  # Scroll
-    "1F5C2",  # Ledger
-    "270D",  # Writing Hand
-    "FE0F",  # Variation Selector-16
-    "202F",  # NARROW NO-BREAK SPACE
-    "00A0",  # NO-BREAK SPACE
-]
-SAFE_EXCLUDED_CHARS = [
-    "£",  # Pound Sign
-    "ç",  # Latin Small Letter C with Cedilla
-    "é",  # Latin Small Letter E with Acute
-    "œ",  # Latin Small Ligature OE
-    "æ",  # Latin Small Letter AE
-    "á",  # Latin Small Letter A with Acute
-    "í",  # Latin Small Letter I with Acute
-    "ñ",  # Latin Small Letter N with Tilde
-    "ö",  # Latin Small Letter O with Diaeresis
-    "ü",  # Latin Small Letter U with Diaeresis
-    "è",  # Latin Small Letter E with Grave
-    "ë",  # Latin Small Letter E with Diaeresis
-    "ï",  # Latin Small Letter I with Diaeresis
-    "ô",  # Latin Small Letter O with Circumflex
-    "ä",  # Latin Small Letter A with Diaeresis
-    "ê",  # Latin Small Letter E with Circumflex
-    "À",  # Latin Capital Letter A with Grave
-    "Ç",  # Latin Capital Letter C with Cedilla
-    "É",  # Latin Capital Letter E with Acute
-    "Ê",  # Latin Capital Letter E with Circumflex
-    "Î",  # Latin Capital Letter I with Circumflex
-    "Ü",  # Latin Capital Letter U with Diaeresis
-    "à",  # Latin Small Letter A with Grave
-    "ó",  # Latin Small Letter O with Acute
-    "ù",  # Latin Small Letter U with Grave
-    "û",  # Latin Small Letter U with Circumflex
-    "―",  # Horizontal Bar
-]
-RARE_REVIEW_CHARS = [
-    "°",  # Degree Sign
-    "´",  # Acute Accent
-    "×",  # Multiplication Sign
-    "ā",  # Latin Small Letter A with Macron
-    "ī",  # Latin Small Letter I with Macron
-    "ū",  # Latin Small Letter U with Macron
-    "Ō",  # Latin Capital Letter O with Macron
-    "ō",  # Latin Small Letter O with Macron
-]
-MOJIBAKE_TRIGGER_CHARS = [
-    "Â",  # Latin Capital Letter A with Circumflex
-    "Ã",  # Latin Capital Letter A with Tilde
-    "â",  # Latin Small Letter A with Circumflex
-]
-DEFAULT_EXCLUDED_CHARS = (
-    SAFE_EXCLUDED_CHARS + RARE_REVIEW_CHARS + MOJIBAKE_TRIGGER_CHARS
-)
-DEFAULT_CHECKS = ["special-characters"]
-DEFAULT_OUTPUT_FORMAT = "human"
-TSV_COLUMNS = [
-    "path",
-    "check",
-    "kind",
-    "severity",
-    "codepoint",
-    "char",
-    "unicode_name",
-    "occurrence_index",
-    "offset",
-    "span_start",
-    "span_end",
-    "context",
-    "classification",
-    "evidence",
-    "message",
-]
-
-BOUNDARY_WINDOW_LINES = 12
-STRUCTURAL_WINDOW_LINES = 40
-
-
-@dataclass(frozen=True)
-class Finding:
-    """One detector finding with stable, testable structure."""
-
-    kind: str
-    severity: str
-    span: tuple[int, int]
-    message: str
-    evidence: Mapping[str, Any]
-
-
-class Detector(Protocol):
-    """Detector interface for corpus survey checks."""
-
-    def run(self, text: str) -> list[Finding]:
-        """Return findings detected in `text`."""
-
-
-class SpecialCharactersDetector:
-    """Classify unusual characters with safe, review, and mojibake buckets."""
-
-    _A_CIRCUMFLEX_MOJIBAKE_RE = re.compile(r"Â(?:[\u00A0-\u00BF]|[^\w\s])")
-    _A_TILDE_MOJIBAKE_RE = re.compile(r"Ã[\u00A0-\u00BF]")
-    _BROKEN_PUNCT_MOJIBAKE_SEQS = {
-        "â€™",
-        "â€œ",
-        "â€\u009d",
-        "â€“",
-        "â€”",
-        "â€¦",
-    }
-
-    def __init__(
-        self,
-        *,
-        allow_smart: bool = True,
-        safe_excluded_chars: Optional[Sequence[str]] = None,
-        rare_review_chars: Optional[Sequence[str]] = None,
-        mojibake_trigger_chars: Optional[Sequence[str]] = None,
-        excluded_chars: Optional[Sequence[str]] = None,
-    ):
-        self.allow_smart = allow_smart
-        legacy_excluded_chars = set(excluded_chars or [])
-        self.safe_excluded_chars = (
-            set(safe_excluded_chars or []) | legacy_excluded_chars
-        )
-        self.rare_review_chars = set(rare_review_chars or [])
-        self.mojibake_trigger_chars = set(mojibake_trigger_chars or [])
-
-    def _is_allowed(self, ch: str) -> bool:
-        if ch.isascii():
-            if ch.isalnum() or ch in " \t\r\n" or ch in ASCII_PUNCT:
-                return True
-        if self.allow_smart and ch in SMART_ALLOWED:
-            return True
-        return False
-
-    def _detect_mojibake_sequences(self, text: str) -> list[Finding]:
-        findings = []
-        pattern_checks = [
-            ("A-circumflex-sequence", self._A_CIRCUMFLEX_MOJIBAKE_RE),
-            ("A-tilde-sequence", self._A_TILDE_MOJIBAKE_RE),
-        ]
-        for pattern_type, pattern in pattern_checks:
-            for match in pattern.finditer(text):
-                findings.append(
-                    Finding(
-                        kind="mojibake-sequence",
-                        severity="error",
-                        span=(match.start(), match.end()),
-                        message="Likely mojibake sequence.",
-                        evidence={
-                            "type": pattern_type,
-                            "sequence": match.group(0),
-                        },
-                    )
-                )
-
-        for sequence in self._BROKEN_PUNCT_MOJIBAKE_SEQS:
-            start = 0
-            while True:
-                index = text.find(sequence, start)
-                if index == -1:
-                    break
-                findings.append(
-                    Finding(
-                        kind="mojibake-sequence",
-                        severity="error",
-                        span=(index, index + len(sequence)),
-                        message="Likely mojibake punctuation sequence.",
-                        evidence={
-                            "type": "broken-punctuation-sequence",
-                            "sequence": sequence,
-                        },
-                    )
-                )
-                start = index + 1
-
-        findings.sort(key=lambda finding: finding.span)
-        return findings
-
-    def run(self, text: str) -> list[Finding]:
-        findings = self._detect_mojibake_sequences(text)
-        mojibake_covered_indexes = set()
-        for finding in findings:
-            for index in range(finding.span[0], finding.span[1]):
-                mojibake_covered_indexes.add(index)
-
-        for idx, ch in enumerate(text):
-            if idx in mojibake_covered_indexes:
-                continue
-            if ch in self.safe_excluded_chars or self._is_allowed(ch):
-                continue
-            if ch in self.rare_review_chars:
-                findings.append(
-                    Finding(
-                        kind="rare-review-character",
-                        severity="info",
-                        span=(idx, idx + 1),
-                        message=f"Uncommon character for review U+{ord(ch):04X}",
-                        evidence={
-                            "character": ch,
-                            "codepoint": f"U+{ord(ch):04X}",
-                            "unicode_name": unicodedata.name(ch, "UNKNOWN"),
-                        },
-                    )
-                )
-                continue
-            if ch in self.mojibake_trigger_chars:
-                continue
-
-            findings.append(
-                Finding(
-                    kind="special-character",
-                    severity="warning",
-                    span=(idx, idx + 1),
-                    message=f"Unexpected character U+{ord(ch):04X}",
-                    evidence={
-                        "character": ch,
-                        "codepoint": f"U+{ord(ch):04X}",
-                        "unicode_name": unicodedata.name(ch, "UNKNOWN"),
-                    },
-                )
-            )
-        return findings
-
-
-def _line_offsets(text: str) -> list[tuple[int, str]]:
-    """Return (start_index, line_text_without_newline) for each line."""
-    offsets = []
-    position = 0
-    for raw_line in text.splitlines(keepends=True):
-        line = raw_line.rstrip("\r\n")
-        offsets.append((position, line))
-        position += len(raw_line)
-
-    if not offsets and text:
-        offsets.append((0, text))
-
-    return offsets
-
-
-def _make_line_finding(
-    *,
-    kind: str,
-    severity: str,
-    line_start: int,
-    line_text: str,
-    message: str,
-    evidence: Mapping[str, Any],
-) -> Finding:
-    """Create a line-based finding with a precise character span."""
-    return Finding(
-        kind=kind,
-        severity=severity,
-        span=(line_start, line_start + len(line_text)),
-        message=message,
-        evidence=evidence,
-    )
-
-
-class StartDetector:
-    """Detect likely non-story preamble content near the story start."""
-
-    _AUTHOR_RE = re.compile(r"^by\s+[A-Za-z][A-Za-z .,'\-]+$", re.IGNORECASE)
-    _EDITOR_NOTE_RE = re.compile(
-        r"^(\[)?(editor'?s note|transcriber'?s note|introduction|preface)\b",
-        re.IGNORECASE,
-    )
-
-    def run(self, text: str) -> list[Finding]:
-        findings = []
-        for line_no, (start, line) in enumerate(_line_offsets(text), start=1):
-            trimmed = line.strip()
-            if not trimmed:
-                continue
-            if line_no > BOUNDARY_WINDOW_LINES:
-                break
-
-            if self._AUTHOR_RE.match(trimmed):
-                findings.append(
-                    _make_line_finding(
-                        kind="start-contamination",
-                        severity="warning",
-                        line_start=start,
-                        line_text=line,
-                        message="Likely author line at start of story.",
-                        evidence={"line": trimmed, "type": "author-line"},
-                    )
-                )
-                continue
-
-            if self._EDITOR_NOTE_RE.match(trimmed):
-                findings.append(
-                    _make_line_finding(
-                        kind="start-contamination",
-                        severity="warning",
-                        line_start=start,
-                        line_text=line,
-                        message="Likely editorial note at start of story.",
-                        evidence={"line": trimmed, "type": "editor-note"},
-                    )
-                )
-                continue
-
-            if line_no <= 3 and self._looks_like_title(trimmed):
-                findings.append(
-                    _make_line_finding(
-                        kind="start-contamination",
-                        severity="warning",
-                        line_start=start,
-                        line_text=line,
-                        message="Likely title heading at story start.",
-                        evidence={"line": trimmed, "type": "title-line"},
-                    )
-                )
-
-        return findings
-
-    @staticmethod
-    def _looks_like_title(line: str) -> bool:
-        if len(line) < 4 or len(line) > 90:
-            return False
-        if line.endswith((".", "!", "?", ";", ":")):
-            return False
-        words = [word for word in re.split(r"\s+", line) if word]
-        if len(words) < 2 or len(words) > 12:
-            return False
-        alpha_words = [word for word in words if any(ch.isalpha() for ch in word)]
-        if len(alpha_words) < 2:
-            return False
-        titleish = sum(word[0].isupper() for word in alpha_words if word[0].isalpha())
-        return titleish >= max(2, len(alpha_words) - 1)
-
-
-class EndDetector:
-    """Detect likely non-story footer content near the story end."""
-
-    _GUTENBERG_RE = re.compile(
-        r"(project gutenberg|gutenberg (ebook|license)|\*\*\* end of)",
-        re.IGNORECASE,
-    )
-
-    def run(self, text: str) -> list[Finding]:
-        lines = _line_offsets(text)
-        if not lines:
-            return []
-
-        findings = []
-        start_index = max(0, len(lines) - BOUNDARY_WINDOW_LINES)
-        for line_no in range(start_index, len(lines)):
-            start, line = lines[line_no]
-            trimmed = line.strip()
-            if not trimmed:
-                continue
-
-            if trimmed.upper() == "THE END":
-                findings.append(
-                    _make_line_finding(
-                        kind="end-contamination",
-                        severity="warning",
-                        line_start=start,
-                        line_text=line,
-                        message="Likely explicit ending marker.",
-                        evidence={"line": trimmed, "type": "the-end"},
-                    )
-                )
-                continue
-
-            if self._GUTENBERG_RE.search(trimmed):
-                findings.append(
-                    _make_line_finding(
-                        kind="end-contamination",
-                        severity="warning",
-                        line_start=start,
-                        line_text=line,
-                        message="Likely Gutenberg footer content.",
-                        evidence={"line": trimmed, "type": "gutenberg-footer"},
-                    )
-                )
-
-        return findings
-
-
-class ChapterHeadingDetector:
-    """Detect chapter-style heading lines embedded in story text."""
-
-    _CHAPTER_RE = re.compile(
-        r"^(chapter|book|part)\s+([ivxlcdm]+|\d+)([\s:.-]+.*)?$", re.IGNORECASE
-    )
-
-    def run(self, text: str) -> list[Finding]:
-        findings = []
-        for line_no, (start, line) in enumerate(_line_offsets(text), start=1):
-            if line_no > STRUCTURAL_WINDOW_LINES:
-                break
-            trimmed = line.strip()
-            if not trimmed or len(trimmed) > 90:
-                continue
-            if self._CHAPTER_RE.match(trimmed):
-                findings.append(
-                    _make_line_finding(
-                        kind="structural-artifact",
-                        severity="warning",
-                        line_start=start,
-                        line_text=line,
-                        message="Likely chapter heading.",
-                        evidence={"line": trimmed, "type": "chapter-heading"},
-                    )
-                )
-        return findings
-
-
-class TocRemnantsDetector:
-    """Detect table-of-contents remnants such as dotted leader lines."""
-
-    _CONTENTS_HEADER_RE = re.compile(r"^(table of contents|contents)$", re.IGNORECASE)
-    _TOC_ENTRY_RE = re.compile(
-        r"^([ivxlcdm\d]+\.?\s+)?[A-Za-z][A-Za-z ,.'\-]+\.{2,}\s*\d+\s*$"
-    )
-
-    def run(self, text: str) -> list[Finding]:
-        lines = _line_offsets(text)[:STRUCTURAL_WINDOW_LINES]
-        findings = []
-        entry_count = 0
-
-        for start, line in lines:
-            trimmed = line.strip()
-            if not trimmed:
-                continue
-
-            if self._CONTENTS_HEADER_RE.match(trimmed):
-                findings.append(
-                    _make_line_finding(
-                        kind="structural-artifact",
-                        severity="warning",
-                        line_start=start,
-                        line_text=line,
-                        message="Likely table-of-contents heading.",
-                        evidence={"line": trimmed, "type": "toc-heading"},
-                    )
-                )
-                continue
-
-            if self._TOC_ENTRY_RE.match(trimmed):
-                entry_count += 1
-                findings.append(
-                    _make_line_finding(
-                        kind="structural-artifact",
-                        severity="warning",
-                        line_start=start,
-                        line_text=line,
-                        message="Likely table-of-contents entry.",
-                        evidence={"line": trimmed, "type": "toc-entry"},
-                    )
-                )
-
-        if entry_count < 2:
-            findings = [
-                finding
-                for finding in findings
-                if finding.evidence.get("type") != "toc-entry"
-            ]
-
-        return findings
-
-
-class SectionBreakDetector:
-    """Detect standalone section-break ornament lines."""
-
-    _SECTION_RE = re.compile(r"^(\*\s*){3,}$|^(\-\s*){3,}$|^~{3,}$|^\*\s\*\s\*$")
-
-    def run(self, text: str) -> list[Finding]:
-        findings = []
-        for start, line in _line_offsets(text):
-            trimmed = line.strip()
-            if not trimmed:
-                continue
-            if self._SECTION_RE.match(trimmed):
-                findings.append(
-                    _make_line_finding(
-                        kind="structural-artifact",
-                        severity="warning",
-                        line_start=start,
-                        line_text=line,
-                        message="Likely section break marker.",
-                        evidence={"line": trimmed, "type": "section-break"},
-                    )
-                )
-        return findings
-
-
-class IllustrationCaptionDetector:
-    """Detect illustration caption lines that are likely non-story artifacts."""
-
-    _CAPTION_RE = re.compile(
-        r"^(\[?illustration\]?[:.]?|figure\s+\d+[:.]|fig\.\s*\d+[:.])",
-        re.IGNORECASE,
-    )
-
-    def run(self, text: str) -> list[Finding]:
-        findings = []
-        for start, line in _line_offsets(text):
-            trimmed = line.strip()
-            if not trimmed or len(trimmed) > 120:
-                continue
-            if self._CAPTION_RE.match(trimmed):
-                findings.append(
-                    _make_line_finding(
-                        kind="structural-artifact",
-                        severity="warning",
-                        line_start=start,
-                        line_text=line,
-                        message="Likely illustration caption.",
-                        evidence={"line": trimmed, "type": "illustration-caption"},
-                    )
-                )
-        return findings
-
-
-def run_detectors(
-    text: str, config: Optional[Mapping[str, Any]] = None
-) -> list[Finding]:
-    """Run configured detectors and return all findings in deterministic order."""
-    resolved = dict(config or {})
-    detectors = resolved.get("detectors")
-
-    if detectors is None:
-        detectors = [
-            SpecialCharactersDetector(
-                safe_excluded_chars=SAFE_EXCLUDED_CHARS,
-                rare_review_chars=RARE_REVIEW_CHARS,
-                mojibake_trigger_chars=MOJIBAKE_TRIGGER_CHARS,
-            ),
-            StartDetector(),
-            EndDetector(),
-            ChapterHeadingDetector(),
-            TocRemnantsDetector(),
-            SectionBreakDetector(),
-            IllustrationCaptionDetector(),
-        ]
-
-    findings = []
-    for detector in detectors:
-        findings.extend(detector.run(text))
-
-    return findings
-
-
-def find_json_files(directories):
-    """Yield JSON files from provided paths in deterministic order."""
-    for directory in directories:
-        path = pathlib.Path(directory)
-        if not path.exists():
-            print(f"warning: directory does not exist: {directory}", file=sys.stderr)
-            continue
-        if path.is_file():
-            if path.suffix == ".json":
-                yield path
-            continue
-        for file_path in sorted(path.rglob("*.json")):
-            if file_path.is_file():
-                yield file_path
-
-
-def run_lcats_display(file_path: pathlib.Path) -> str:
-    """Render one corpus JSON file using the same formatter as `lcats display`."""
-    with file_path.open("r", encoding="utf-8") as json_file:
-        data = json.load(json_file)
-
-    rendered = lcats.inspect.format_story_json(data, max_body_chars=None, width=80)
-    return f"{rendered}\n"
-
-
-def run_special_characters_check(
-    displayed_text: str,
-    extract_script: str,
-    allow_smart: bool,
-    allowlist_config: str,
-    excluded_codepoints,
-    excluded_chars,
-    context: int,
-    nocontext: bool,
-    name_width: int,
-    header: bool,
-) -> str:
-    """Run the special-character extractor and return its TSV output."""
-    cmd = [extract_script]
-
-    if allow_smart:
-        cmd.append("--allow-smart")
-    if allowlist_config:
-        cmd.append(f"--allowlist-config={allowlist_config}")
-    if excluded_codepoints:
-        cmd.append("--exclude-codepoint=" + ",".join(excluded_codepoints))
-    if excluded_chars:
-        cmd.append("--exclude-char=" + ",".join(excluded_chars))
-    if nocontext:
-        cmd.append("--nocontext")
-    else:
-        cmd.append(f"--context={context}")
-    if name_width > 0:
-        cmd.append(f"--name-width={name_width}")
-    if header:
-        cmd.append("--header")
-
-    result = subprocess.run(
-        cmd,
-        input=displayed_text,
-        capture_output=True,
-        text=True,
-        encoding="utf-8",
-    )
-    if result.returncode not in (0, 141):
-        raise RuntimeError(
-            "special-character check failed:\n"
-            f"stdout:\n{result.stdout}\n"
-            f"stderr:\n{result.stderr}"
-        )
-    return result.stdout.strip()
-
-
-def run_boundary_contamination_check(displayed_text: str) -> str:
-    """Run boundary contamination detectors and emit TSV rows."""
-    detectors = [StartDetector(), EndDetector()]
-    findings = run_detectors(displayed_text, config={"detectors": detectors})
-    lines = []
-    for finding in findings:
-        lines.append(
-            "\t".join(
-                [
-                    finding.kind,
-                    finding.severity,
-                    str(finding.span[0]),
-                    str(finding.span[1]),
-                    finding.message,
-                ]
-            )
-        )
-    return "\n".join(lines)
-
-
-def _empty_tsv_row() -> dict[str, str]:
-    """Return a blank TSV row with all stable fields present."""
-    return {column: "" for column in TSV_COLUMNS}
-
-
-def _severity_from_classification(classification: str) -> str:
-    """Map special-character classification to severity."""
-    lowered = classification.lower()
-    if "mojibake" in lowered:
-        return "error"
-    if "rare" in lowered:
-        return "info"
-    if lowered:
-        return "warning"
-    return "warning"
-
-
-def parse_special_character_rows(
-    tsv_output: str, file_path: pathlib.Path
-) -> list[dict[str, str]]:
-    """Parse extractor TSV output into stable report rows."""
-    if not tsv_output.strip():
-        return []
-
-    rows = []
-    parsed_rows = list(csv.reader(tsv_output.splitlines(), delimiter="\t"))
-    if parsed_rows and parsed_rows[0] and parsed_rows[0][0] == "codepoint":
-        parsed_rows = parsed_rows[1:]
-
-    for parts in parsed_rows:
-        if not parts:
-            continue
-        padded = parts + [""] * (8 - len(parts))
-        row = _empty_tsv_row()
-        row.update(
-            {
-                "path": str(file_path),
-                "check": "special-characters",
-                "kind": "special-character",
-                "severity": _severity_from_classification(padded[6]),
-                "codepoint": padded[0],
-                "char": padded[1],
-                "unicode_name": padded[2],
-                "occurrence_index": padded[3],
-                "offset": padded[4],
-                "context": padded[5],
-                "classification": padded[6],
-                "evidence": padded[7],
-                "message": "Special character finding.",
-            }
-        )
-        rows.append(row)
-
-    return rows
-
-
-def _finding_to_row(
-    file_path: pathlib.Path, check_name: str, finding: Finding
-) -> dict[str, str]:
-    """Convert one Finding into a stable TSV row."""
-    row = _empty_tsv_row()
-    row.update(
-        {
-            "path": str(file_path),
-            "check": check_name,
-            "kind": finding.kind,
-            "severity": finding.severity,
-            "span_start": str(finding.span[0]),
-            "span_end": str(finding.span[1]),
-            "evidence": json.dumps(dict(finding.evidence), ensure_ascii=False),
-            "message": finding.message,
-        }
-    )
-    return row
-
-
-def _clean_row(file_path: pathlib.Path) -> dict[str, str]:
-    """Build a summary row for files with no findings."""
-    row = _empty_tsv_row()
-    row.update(
-        {
-            "path": str(file_path),
-            "check": "summary",
-            "kind": "clean",
-            "severity": "info",
-            "message": "No findings.",
-        }
-    )
-    return row
-
-
-def survey_file(file_path: pathlib.Path, args) -> list[dict[str, str]]:
-    """Run enabled checks on a single corpus file and return report rows."""
-    displayed_text = run_lcats_display(file_path)
-    rows = []
-
-    for check_name in args.check_for:
-        if check_name == "special-characters":
-            output = run_special_characters_check(
-                displayed_text=displayed_text,
-                extract_script=args.extract_script,
-                allow_smart=args.allow_smart,
-                allowlist_config=args.allowlist_config,
-                excluded_codepoints=args.exclude_codepoint,
-                excluded_chars=args.exclude_char,
-                context=args.context,
-                nocontext=args.nocontext,
-                name_width=args.name_width,
-                header=False,
-            )
-            rows.extend(parse_special_character_rows(output, file_path))
-        elif check_name == "boundary-contamination":
-            boundary_findings = run_detectors(
-                displayed_text, config={"detectors": [StartDetector(), EndDetector()]}
-            )
-            rows.extend(
-                _finding_to_row(file_path, check_name, finding)
-                for finding in boundary_findings
-            )
-        else:
-            raise ValueError(f"Unsupported check: {check_name}")
-
-    return rows
-
-
-def parse_csv_args(values):
-    """Split repeatable comma-separated CLI values into a flat list."""
-    items = []
-    for value in values or []:
-        for item in value.split(","):
-            item = item.strip()
-            if item:
-                items.append(item)
-    return items
-
-
-def build_parser():
-    """Build and return the CLI parser for corpus quality checks."""
-    parser = argparse.ArgumentParser(
-        description="Survey LCATS corpora files for issues."
-    )
-    parser.add_argument(
-        "directories",
-        nargs="*",
-        default=["data/"],
-        help="Directories or files to survey.",
-    )
-    parser.add_argument(
-        "--check-for",
-        action="append",
-        default=[],
-        help=(
-            "Check(s) to run. Repeatable or comma-separated. "
-            "Currently supported: special-characters,boundary-contamination"
-        ),
-    )
-    parser.add_argument(
-        "--print-clean-filenames",
-        action="store_true",
-        help="Print filenames even when no issues are found.",
-    )
-    parser.add_argument(
-        "--extract-script",
-        default="scripts/utils/extract_special_chars.py",
-        help="Path to the extract_special_chars.py script.",
-    )
-    parser.add_argument(
-        "--allowlist-config",
-        default="",
-        help="Optional JSON config path for allowed characters/codepoints.",
-    )
-
-    parser.add_argument(
-        "--allow-smart",
-        dest="allow_smart",
-        action="store_true",
-        help="Allow common smart punctuation (default: enabled).",
-    )
-    parser.add_argument(
-        "--no-allow-smart",
-        dest="allow_smart",
-        action="store_false",
-        help="Disable smart punctuation allowance.",
-    )
-    parser.set_defaults(allow_smart=True)
-
-    parser.add_argument(
-        "--context",
-        type=int,
-        default=10,
-        help="Number of left/right context characters to include (default: 10).",
-    )
-    parser.add_argument(
-        "--nocontext",
-        action="store_true",
-        help="Convenience flag equivalent to --context=0.",
-    )
-    parser.add_argument(
-        "--name-width",
-        type=int,
-        default=0,
-        help="Optional max width for Unicode name (0 means no truncation).",
-    )
-    parser.add_argument(
-        "--header",
-        dest="header",
-        action="store_true",
-        help="Emit a TSV header row (default in TSV mode).",
-    )
-    parser.add_argument(
-        "--no-header",
-        dest="header",
-        action="store_false",
-        help="Do not emit a TSV header row.",
-    )
-    parser.set_defaults(header=None)
-    parser.add_argument(
-        "--format",
-        choices=["human", "tsv"],
-        default=DEFAULT_OUTPUT_FORMAT,
-        help="Output format (human or machine-readable TSV).",
-    )
-    parser.add_argument(
-        "--output",
-        default="",
-        help="Optional output file path (default: stdout).",
-    )
-    parser.add_argument(
-        "--progress",
-        dest="progress",
-        action="store_true",
-        help="Force-enable the progress bar.",
-    )
-    parser.add_argument(
-        "--no-progress",
-        dest="progress",
-        action="store_false",
-        help="Disable the progress bar.",
-    )
-    parser.set_defaults(progress=None)
-
-    parser.add_argument(
-        "--exclude-codepoint",
-        action="append",
-        default=[],
-        help=(
-            "Exclude Unicode codepoints from the special-characters check. "
-            "Repeatable or comma-separated."
-        ),
-    )
-    parser.add_argument(
-        "--exclude-char",
-        action="append",
-        default=[],
-        help=(
-            "Exclude literal characters from the special-characters check. "
-            "Repeatable or comma-separated."
-        ),
-    )
-    return parser
-
-
-def _show_progress(progress_arg: Optional[bool]) -> bool:
-    """Resolve progress-bar visibility from CLI choice and TTY defaults."""
-    if progress_arg is not None:
-        return progress_arg
-    return sys.stderr.isatty()
-
-
-def _write_human_rows(
-    output_stream, file_path: pathlib.Path, rows: Sequence[Mapping[str, str]]
-) -> None:
-    """Write human-readable findings for one file."""
-    print(str(file_path), file=output_stream)
-    for row in rows:
-        check_name = row.get("check", "")
-        severity = row.get("severity", "")
-        message = row.get("message", "")
-        codepoint = row.get("codepoint", "")
-        character = row.get("char", "")
-        span_start = row.get("span_start", "")
-        span_end = row.get("span_end", "")
-
-        details = []
-        if codepoint:
-            details.append(codepoint)
-        if character:
-            details.append(repr(character))
-        if span_start or span_end:
-            details.append(f"span={span_start}:{span_end}")
-        suffix = f" ({', '.join(details)})" if details else ""
-        print(
-            f"  [{check_name}] {severity}: {message}{suffix}",
-            file=output_stream,
-        )
-
-
-def main(argv: Optional[Sequence[str]] = None) -> int:
-    """Run the corpus quality checker CLI."""
-    parser = build_parser()
-    args = parser.parse_args(argv)
-
-    if args.context < 0:
-        parser.error("--context must be >= 0")
-    if args.name_width < 0:
-        parser.error("--name-width must be >= 0")
-
-    args.check_for = parse_csv_args(args.check_for) or list(DEFAULT_CHECKS)
-
-    user_excluded_codepoints = parse_csv_args(args.exclude_codepoint)
-    args.exclude_codepoint = (
-        list(DEFAULT_EXCLUDED_CODEPOINTS) + user_excluded_codepoints
-    )
-
-    user_excluded_chars = parse_csv_args(args.exclude_char)
-    args.exclude_char = list(DEFAULT_EXCLUDED_CHARS) + user_excluded_chars
-
-    had_findings = False
-
+"""Compatibility wrapper for corpus survey APIs."""
+
+from lcats.analysis.corpus import cli
+from lcats.analysis.corpus import discovery
+from lcats.analysis.corpus import output
+from lcats.analysis.corpus import qa
+from lcats.analysis.corpus.detectors import boundary
+from lcats.analysis.corpus.detectors import structural
+from lcats.analysis.corpus.detectors import unicode
+from lcats.analysis.corpus import models
+
+SMART_ALLOWED = unicode.SMART_ALLOWED
+ASCII_PUNCT = unicode.ASCII_PUNCT
+DEFAULT_EXCLUDED_CODEPOINTS = unicode.DEFAULT_EXCLUDED_CODEPOINTS
+SAFE_EXCLUDED_CHARS = unicode.SAFE_EXCLUDED_CHARS
+RARE_REVIEW_CHARS = unicode.RARE_REVIEW_CHARS
+MOJIBAKE_TRIGGER_CHARS = unicode.MOJIBAKE_TRIGGER_CHARS
+DEFAULT_EXCLUDED_CHARS = unicode.DEFAULT_EXCLUDED_CHARS
+DEFAULT_CHECKS = qa.DEFAULT_CHECKS
+DEFAULT_OUTPUT_FORMAT = output.DEFAULT_OUTPUT_FORMAT
+TSV_COLUMNS = output.TSV_COLUMNS
+BOUNDARY_WINDOW_LINES = boundary.BOUNDARY_WINDOW_LINES
+STRUCTURAL_WINDOW_LINES = structural.STRUCTURAL_WINDOW_LINES
+
+Finding = models.Finding
+
+SpecialCharactersDetector = unicode.SpecialCharactersDetector
+StartDetector = boundary.StartDetector
+EndDetector = boundary.EndDetector
+ChapterHeadingDetector = structural.ChapterHeadingDetector
+TocRemnantsDetector = structural.TocRemnantsDetector
+SectionBreakDetector = structural.SectionBreakDetector
+IllustrationCaptionDetector = structural.IllustrationCaptionDetector
+
+run_detectors = qa.run_detectors
+find_json_files = discovery.find_json_files
+run_lcats_display = cli.run_lcats_display
+run_special_characters_check = cli.run_special_characters_check
+parse_special_character_rows = output.parse_special_character_rows
+survey_file = cli.survey_file
+parse_csv_args = cli.parse_csv_args
+build_parser = cli.build_survey_parser
+subprocess = cli.subprocess
+lcats = cli.lcats
+
+
+def main(argv=None):
+    """Run survey CLI via compatibility wrapper."""
+    original_find_json_files = cli.discovery.find_json_files
+    original_survey_file = cli.survey_file
+    original_tqdm = cli.tqdm
+    original_sys = cli.sys
     try:
-        files_found = list(find_json_files(args.directories))
-        output_stream = sys.stdout
-        if args.output:
-            output_stream = pathlib.Path(args.output).open(
-                "w", encoding="utf-8", newline=""
-            )
+        cli.discovery.find_json_files = find_json_files
+        cli.survey_file = survey_file
+        cli.tqdm = tqdm
+        cli.sys = sys
+        return cli.run_survey(argv)
+    finally:
+        cli.discovery.find_json_files = original_find_json_files
+        cli.survey_file = original_survey_file
+        cli.tqdm = original_tqdm
+        cli.sys = original_sys
 
-        header_enabled = args.header
-        if header_enabled is None:
-            header_enabled = args.format == "tsv"
 
-        tsv_writer = None
-        if args.format == "tsv":
-            tsv_writer = csv.DictWriter(
-                output_stream,
-                fieldnames=TSV_COLUMNS,
-                delimiter="\t",
-                extrasaction="ignore",
-            )
-            if header_enabled:
-                tsv_writer.writeheader()
-
-        for file_path in tqdm.tqdm(
-            files_found, disable=not _show_progress(args.progress)
-        ):
-            rows = survey_file(file_path, args)
-            if rows:
-                had_findings = True
-                if args.format == "tsv":
-                    for row in rows:
-                        tsv_writer.writerow(row)
-                else:
-                    _write_human_rows(output_stream, file_path, rows)
-            elif args.print_clean_filenames:
-                if args.format == "tsv":
-                    tsv_writer.writerow(_clean_row(file_path))
-                else:
-                    print(f"{file_path} [clean]", file=output_stream)
-
-        if output_stream is not sys.stdout:
-            output_stream.close()
-        return 1 if had_findings else 0
-
-    except BrokenPipeError:
-        return 141
-    except Exception as exc:
-        print(f"error: {exc}", file=sys.stderr)
-        return 2
+_clean_row = output.clean_row
+_show_progress = cli._show_progress
+sys = cli.sys
+tqdm = cli.tqdm
